viz_agent/pyproject.toml
[project]
name = "viz-agent"
version = "0.1.0"
description = "Agente de visualización con Gemini AI para Gen BI"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Nahuel Agustin Carro", email = "nahuelagustincarro@gmail.com"}
]

dependencies = [
    # Data & Visualization
    "pandas>=2.2.0",
    "plotly>=5.18.0",
    "kaleido>=0.2.1",
    
    # AI/LLM - Nueva API de Google GenAI
    "google-genai>=0.3.0",
    
    # Utilities
    "pydantic>=2.5.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    
    # Linting & Type checking
    "ruff>=0.1.9",
    "mypy>=1.7.1",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --cov=viz_agent --cov-report=html --cov-report=term"
pythonpath = ["src"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/viz_agent"]
